DataPipeline/reference code/requirements.txt
requests>=2.25.0
PyYAML>=6.0
pandas>=1.3.0 